tests/functional/requirements.txt
pytest==6.2.5
pytest-asyncio==0.12.0
aiohttp==3.7.2

elasticsearch==8.10.1
redis==4.0.0

pydantic==2.4.2
pydantic-settings==2.0.3

python-dotenv==1.0.0
backoff==2.2.1
